Client and Host Seperated
The program now can run as a client if the user specifies Client in the
arguments. Host and Client are in separate methods due to differences
in socket connection handling. This has been tested and connections
were able to be established between a client and host with no warning
or errors.